Conflicts:
	quantlib/market/market.py

+++ quantlib/instruments/payoffs.pyx
@@ -51,7 +51,7 @@ cdef class PlainVanillaPayoff(Payoff):
 
     option_type: int or str
         The type of option, can be either Call or Put
-    strike: float
+    strike: double
         The strike value
     from_qlpayoff: bool, optional
         For internal use only
@@ -64,7 +64,7 @@ cdef class PlainVanillaPayoff(Payoff):
         Read-only property that returns a PlainVanillaPayoff instance
     """
 
-    def __init__(self, option_type, float strike, from_qlpayoff=False):
+    def __init__(self, option_type, double strike, from_qlpayoff=False):
 
         if isinstance(option_type, basestring):
             option_type = str_to_option_type(option_type)
