+++ quantlib/quotes.pyx
@@ -30,7 +30,7 @@ cdef class Quote:
 
 cdef class SimpleQuote(Quote):
 
-    def __init__(self, float value=0.0):
+    def __init__(self, double value=0.0):
         self._thisptr = new shared_ptr[_qt.Quote](new _qt.SimpleQuote(value))
 
     def __str__(self):
@@ -43,5 +43,5 @@ cdef class SimpleQuote(Quote):
             else:
                 return None
 
-        def __set__(self, float value):
+        def __set__(self, double value):
             (<_qt.SimpleQuote*>self._thisptr.get()).setValue(value)
